infra/ansible/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
ansible==2.3.0.0
asn1crypto==0.22.0        # via cryptography
cffi==1.10.0              # via cryptography
cryptography==1.8.1       # via paramiko
enum34==1.1.6             # via cryptography
idna==2.5                 # via cryptography
ipaddress==1.0.18         # via cryptography
jinja2==2.9.6             # via ansible
markupsafe==1.0           # via jinja2
packaging==16.8           # via cryptography
paramiko==2.1.2           # via ansible
pyasn1==0.2.3             # via paramiko
pycparser==2.17           # via cffi
pycrypto==2.6.1           # via ansible
pyparsing==2.2.0          # via packaging
pyyaml==3.12
six==1.10.0               # via cryptography, packaging
